move getBaseUrl to server component
